Reformat to 100 char width, normalize safety docs
