+++ rkyv/src/ser/mod.rs
@@ -13,12 +13,10 @@ use core::{mem, slice};
 /// A byte sink that knows where it is.
 ///
 /// A type that is [`io::Write`](std::io::Write) can be wrapped in a
-/// [`WriteSerializer`](serializers::WriteSerializer) to equip it with
-/// `Serializer`.
+/// [`WriteSerializer`](serializers::WriteSerializer) to equip it with `Serializer`.
 ///
-/// It's important that the memory for archived objects is properly aligned
-/// before attempting to read objects out of it; use the
-/// [`Aligned`](crate::Aligned) wrapper if it's appropriate.
+/// It's important that the memory for archived objects is properly aligned before attempting to
+/// read objects out of it; use the [`Aligned`](crate::Aligned) wrapper if it's appropriate.
 pub trait Serializer: Fallible {
     /// Returns the current position of the serializer.
     fn pos(&self) -> usize;
@@ -46,8 +44,7 @@ pub trait Serializer: Fallible {
         Ok(self.pos())
     }
 
-    /// Aligns the position of the serializer to be suitable to write the given
-    /// type.
+    /// Aligns the position of the serializer to be suitable to write the given type.
     #[inline]
     fn align_for<T>(&mut self) -> Result<usize, Self::Error> {
         self.align(mem::align_of::<T>())
@@ -59,8 +56,8 @@ pub trait Serializer: Fallible {
     ///
     /// # Safety
     ///
-    /// This is only safe to call when the serializer is already aligned for the
-    /// archived version of the given type.
+    /// The caller must guarantee that the serializer is aligned for the archived version of the
+    /// given type.
     unsafe fn resolve_aligned<T: Archive + ?Sized>(
         &mut self,
         value: &T,
@@ -82,15 +79,14 @@ pub trait Serializer: Fallible {
         unsafe { self.resolve_aligned(value, resolver) }
     }
 
-    /// Resolves the given reference with its resolver and writes the archived
-    /// reference.
+    /// Resolves the given reference with its resolver and writes the archived reference.
     ///
     /// Returns the position of the written archived reference.
     ///
     /// # Safety
     ///
-    /// This is only safe to call when the serializer is already aligned for the
-    /// archived reference of the given type.
+    /// The caller must guarantee that the serializer is aligned for the archived reference of the
+    /// given type.
     unsafe fn resolve_unsized_aligned<T: ArchiveUnsized + ?Sized>(
         &mut self,
         value: &T,
@@ -106,8 +102,7 @@ pub trait Serializer: Fallible {
         Ok(from)
     }
 
-    /// Archives a reference to the given object and returns the position it was
-    /// archived at.
+    /// Archives a reference to the given object and returns the position it was archived at.
     fn serialize_unsized_value<T: SerializeUnsized<Self> + ?Sized>(
         &mut self,
         value: &T,
@@ -124,8 +119,8 @@ pub trait SeekSerializer: Serializer {
     /// Seeks the serializer to the given absolute position.
     fn seek(&mut self, pos: usize) -> Result<(), Self::Error>;
 
-    /// Archives the given value at the nearest available position. If the
-    /// serializer is already aligned, it will archive it at the current position.
+    /// Archives the given value at the nearest available position. If the serializer is already
+    /// aligned, it will archive it at the current position.
     fn serialize_front<T: Serialize<Self>>(&mut self, value: &T) -> Result<usize, Self::Error> {
         self.align_for::<T::Archived>()?;
         let pos = self.pos();
@@ -138,9 +133,8 @@ pub trait SeekSerializer: Serializer {
         Ok(pos)
     }
 
-    /// Archives a reference to the given value at the nearest available
-    /// position. If the serializer is already aligned, it will archive it at the
-    /// current position.
+    /// Archives a reference to the given value at the nearest available position. If the serializer
+    /// is already aligned, it will archive it at the current position.
     fn serialize_unsized_front<T: SerializeUnsized<Self> + ?Sized>(
         &mut self,
         value: &T,
@@ -163,9 +157,8 @@ pub trait SeekSerializer: Serializer {
 ///
 /// This serializer is required by shared pointers to serialize.
 pub trait SharedSerializer: Serializer {
-    /// Archives the given shared value and returns its position. If the value
-    /// has already been serialized then it returns the position of the
-    /// previously serialized value.
+    /// Archives the given shared value and returns its position. If the value has already been
+    /// serialized then it returns the position of the previously serialized value.
     fn serialize_shared<T: SerializeUnsized<Self> + ?Sized>(
         &mut self,
         value: &T,
