+++ rkyv/src/validation.rs
@@ -16,14 +16,14 @@ use std::{collections::HashMap, error::Error};
 impl RawRelPtr {
     /// Checks the bytes of the given raw relative pointer.
     ///
-    /// This is done rather than implementing `CheckBytes` to force users to
-    /// manually write their `CheckBytes` implementation since they need to also
-    /// provide the ownership model of their memory.
+    /// This is done rather than implementing `CheckBytes` to force users to manually write their
+    /// `CheckBytes` implementation since they need to also provide the ownership model of their
+    /// memory.
     ///
     /// # Safety
     ///
-    /// The given pointer must be aligned and point to enough bytes to represent
-    /// a `RawRelPtr`.
+    /// The caller must guarantee that the given pointer is aligned and points to enough bytes to
+    /// represent a `RawRelPtr`.
     #[inline]
     pub unsafe fn manual_check_bytes<'a, C: Fallible + ?Sized>(
         value: *const RawRelPtr,
@@ -39,14 +39,14 @@ impl RawRelPtr {
 impl<T: ArchivePointee + ?Sized> RelPtr<T> {
     /// Checks the bytes of the given relative pointer.
     ///
-    /// This is done rather than implementing `CheckBytes` to force users to
-    /// manually write their `CheckBytes` implementation since they need to also
-    /// provide the ownership model of their memory.
+    /// This is done rather than implementing `CheckBytes` to force users to manually write their
+    /// `CheckBytes` implementation since they need to also provide the ownership model of their
+    /// memory.
     ///
     /// # Safety
     ///
-    /// The given pointer must be aligned and point to enough bytes to represent
-    /// a `RelPtr<T>`.
+    /// The caller must guarantee that the given pointer is aligned and points to enough bytes to
+    /// represent a `RelPtr<T>`.
     #[inline]
     pub unsafe fn manual_check_bytes<'a, C: Fallible + ?Sized>(
         value: *const RelPtr<T>,
@@ -183,7 +183,7 @@ pub trait ArchiveBoundsContext: Fallible {
     ///
     /// # Safety
     ///
-    /// The base pointer must be inside the archive for this context.
+    /// The caller must guarantee that the base pointer is inside the archive for this context.
     unsafe fn check_rel_ptr(
         &mut self,
         base: *const u8,
@@ -194,7 +194,7 @@ pub trait ArchiveBoundsContext: Fallible {
     ///
     /// # Safety
     ///
-    /// The pointer must be inside the archive for this context.
+    /// The caller must guarantee that the pointer is inside the archive for this context.
     unsafe fn bounds_check_ptr(
         &mut self,
         ptr: *const u8,
@@ -343,24 +343,23 @@ impl<E: Error + 'static> Error for ArchiveMemoryError<E> {
 
 /// A context that can validate archive memory.
 ///
-/// When implementing archivable containers, an archived type may point to some
-/// bytes elsewhere in the archive using a [`RelPtr`]. Before checking those
-/// bytes, they must be claimed in the context. This prevents infinite-loop
-/// attacks by malicious actors by ensuring that each block of memory has one
-/// and only one owner.
+/// When implementing archivable containers, an archived type may point to some bytes elsewhere in
+/// the archive using a [`RelPtr`]. Before checking those bytes, they must be claimed in the
+/// context. This prevents infinite-loop attacks by malicious actors by ensuring that each block of
+/// memory has one and only one owner.
 pub trait ArchiveMemoryContext: Fallible {
     /// Claims `count` bytes located `offset` bytes away from `base`.
     ///
     /// # Safety
     ///
-    /// `base` must be inside the archive this context was created for.
+    /// The caller must guarantee that `base` is inside the archive this context was created for.
     unsafe fn claim_bytes(&mut self, start: *const u8, len: usize) -> Result<(), Self::Error>;
 
     /// Claims the memory at the given location as the given type.
     ///
     /// # Safety
     ///
-    /// `ptr` must be inside the archive this context was created for.
+    /// The caller must guarantee that `ptr` is inside the archive this context was created for.
     unsafe fn claim_owned_ptr<T: ArchivePointee + ?Sized>(
         &mut self,
         ptr: *const T,
@@ -535,7 +534,7 @@ pub trait SharedArchiveContext: Fallible {
     ///
     /// # Safety
     ///
-    /// `base` must be inside the archive this context was created for.
+    /// The caller must guarantee that `base` is inside the archive this context was created for.
     unsafe fn claim_shared_bytes(
         &mut self,
         start: *const u8,
@@ -545,8 +544,7 @@ pub trait SharedArchiveContext: Fallible {
 
     /// Claims the memory referenced by the given relative pointer.
     ///
-    /// If the pointer needs to be checked, returns `Some` with the pointer to
-    /// check.
+    /// If the pointer needs to be checked, returns `Some` with the pointer to check.
     fn claim_shared_ptr<T: ArchivePointee + CheckBytes<Self> + ?Sized>(
         &mut self,
         rel_ptr: &RelPtr<T>,
@@ -691,8 +689,7 @@ pub type DefaultArchiveValidator = SharedArchiveValidator<ArchiveValidator<Archi
 pub type CheckTypeError<T, C> =
     CheckArchiveError<<T as CheckBytes<C>>::Error, <C as Fallible>::Error>;
 
-/// Checks the given archive at the given position for an archived version of
-/// the given type.
+/// Checks the given archive at the given position for an archived version of the given type.
 ///
 /// This is a safe alternative to [`archived_value`](crate::archived_value) for types that implement
 /// `CheckBytes`.
@@ -739,8 +736,7 @@ where
     check_archived_value_with_context::<T, DefaultArchiveValidator>(buf, pos, &mut validator)
 }
 
-/// Checks the given archive at the given position for an archived version of
-/// the given type.
+/// Checks the given archive at the given position for an archived version of the given type.
 ///
 /// This is a safe alternative to [`archived_value`](crate::archived_value) for types that implement
 /// `CheckBytes`.
